requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
boto3==1.34.0
python-dotenv==1.0.0
plotly==5.17.0
Werkzeug==2.3.7
gunicorn==21.2.0 